質問に関するDAOクラスを作成。また、命名変更
